# specs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Servicing API",
    "description": "API to get service details about customers"
  },
  "servers": [
    {
      "url": "host"
    }
  ],
...
  "paths": {
...
    "/customer_accounts/{id}": {
      "get": {
        "tags": [
          "customer_accounts"
        ],
        "summary": "Retrieves customer information by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a customer account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/customer_account"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "patch": {
        "tags": [
          "customer_accounts"
        ],
        "summary": "Updates customers email and/or primary phone number",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/customer_account_patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a customer account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/customer_account"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/customer_accounts/{id}/notes": {
      "post": {
        "tags": [
          "customer_accounts",
          "notes"
        ],
        "summary": "Adds a note to the customer account.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to which note is added.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/customer_account_add_note_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a success flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Is_Success": {
                      "type": "boolean"
                    }
                  }
                },
                "example": {
                  "Is_Success": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
    "/customer_accounts/{id}/notification_options": {
      "get": {
        "tags": [
          "customer_accounts",
          "notifications"
        ],
        "summary": "Retrieves notification options for a customer account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to fetch notification options for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "description": "type of notification option to search for",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Bill",
                "Contract_Expiration",
                "Payment_Due",
                "Payment_Received"
              ]
            }
          },
          {
            "name": "Status",
            "in": "query",
            "description": "status of notification option",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Suspended",
                "Canceled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns list of notification options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Id": {
                        "type": "string",
                        "pattern": "^[a-fA-F0-9]{24}$",
                        "readOnly": true
                      },
                      "Date_Created": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                      },
                      "Date_Last_Modified": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                      },
                      "Type": {
                        "type": "string",
                        "enum": [
                          "Bill",
                          "Contract_Expiration",
                          "Payment_Due",
                          "Payment_Received"
                        ]
                      },
                      "Status": {
                        "type": "string",
                        "enum": [
                          "Active",
                          "Suspended",
                          "Canceled"
                        ]
                      },
                      "Paperless": {
                        "type": "boolean"
                      },
                      "Preferred_Contact_Method": {
                        "type": "string",
                        "enum": [
                          "Email",
                          "Phone"
                        ]
                      },
                      "Email": {
                        "type": "string",
                        "format": "email"
                      },
                      "Phone_Number": {
                        "type": "object",
                        "properties": {
                          "Type": {
                            "type": "string",
                            "enum": [
                              "Landline",
                              "Mobile"
                            ]
                          },
                          "Area_Code": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 3,
                            "pattern": "^\\d{3}$"
                          },
                          "Number": {
                            "type": "string",
                            "minLength": 7,
                            "maxLength": 7,
                            "pattern": "^\\d{7}$"
                          },
                          "Extension": {
                            "type": "string"
                          },
                          "Agree_To_Marketing": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "Type",
                          "Area_Code",
                          "Number"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "customer_accounts",
          "notifications"
        ],
        "summary": "Set notification option",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to set notification option for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Type": {
                    "type": "string",
                    "enum": [
                      "Bill",
                      "Contract_Expiration",
                      "Payment_Due",
                      "Payment_Received"
                    ]
                  },
                  "Paperless": {
                    "type": "boolean"
                  },
                  "Preferred_Contact_Method": {
                    "type": "string",
                    "enum": [
                      "Email",
                      "Phone"
                    ]
                  },
                  "Email": {
                    "type": "string",
                    "format": "email"
                  },
                  "Phone_Number": {
                    "type": "object",
                    "properties": {
                      "Type": {
                        "type": "string",
                        "enum": [
                          "Landline",
                          "Mobile"
                        ]
                      },
                      "Area_Code": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 3,
                        "pattern": "^\\d{3}$"
                      },
                      "Number": {
                        "type": "string",
                        "minLength": 7,
                        "maxLength": 7,
                        "pattern": "^\\d{7}$"
                      },
                      "Extension": {
                        "type": "string"
                      },
                      "Agree_To_Marketing": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "Type",
                      "Area_Code",
                      "Number"
                    ]
                  }
                },
                "required": [
                  "Type",
                  "Paperless",
                  "Preferred_Contact_Method"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a notification option",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string",
                      "pattern": "^[a-fA-F0-9]{24}$",
                      "readOnly": true
                    },
                    "Date_Created": {
                      "type": "string",
                      "format": "date-time",
                      "readOnly": true
                    },
                    "Date_Last_Modified": {
                      "type": "string",
                      "format": "date-time",
                      "readOnly": true
                    },
                    "Type": {
                      "type": "string",
                      "enum": [
                        "Bill",
                        "Contract_Expiration",
                        "Payment_Due",
                        "Payment_Received"
                      ]
                    },
                    "Status": {
                      "type": "string",
                      "enum": [
                        "Active",
                        "Suspended",
                        "Canceled"
                      ]
                    },
                    "Paperless": {
                      "type": "boolean"
                    },
                    "Preferred_Contact_Method": {
                      "type": "string",
                      "enum": [
                        "Email",
                        "Phone"
                      ]
                    },
                    "Email": {
                      "type": "string",
                      "format": "email"
                    },
                    "Phone_Number": {
                      "type": "object",
                      "properties": {
                        "Type": {
                          "type": "string",
                          "enum": [
                            "Landline",
                            "Mobile"
                          ]
                        },
                        "Area_Code": {
                          "type": "string",
                          "minLength": 3,
                          "maxLength": 3,
                          "pattern": "^\\d{3}$"
                        },
                        "Number": {
                          "type": "string",
                          "minLength": 7,
                          "maxLength": 7,
                          "pattern": "^\\d{7}$"
                        },
                        "Extension": {
                          "type": "string"
                        },
                        "Agree_To_Marketing": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "Type",
                        "Area_Code",
                        "Number"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer Account not found"
          }
        }
      }
    },
...
    "/customer_accounts/{id}/pay_methods": {
      "get": {
        "tags": [
          "customer_accounts",
          "payments"
        ],
        "summary": "Retrieves payment methods for a customer by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer to fetch payment methods for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "active status of pay method",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a list of pay methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/pay_method"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No payment methods not found"
          }
        }
      },
      "post": {
        "tags": [
          "customer_accounts",
          "payments"
        ],
        "summary": "Add payment account for the customer account specified",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to add paymethod for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Token": {
                    "description": "temporary use token issued from the payment vendor",
                    "type": "string",
                    "pattern": "^ott_"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a pay method",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "integer"
                    },
                    "Active": {
                      "type": "boolean"
                    },
                    "Type": {
                      "type": "string",
                      "enum": [
                        "CreditCard",
                        "eCheck",
                        "Unknown"
                      ]
                    },
                    "CreditCard": {
                      "type": "object",
                      "properties": {
                        "AccountNumber": {
                          "type": "string"
                        },
                        "AccountHolder": {
                          "type": "string"
                        },
                        "ExpirationYear": {
                          "type": "integer"
                        },
                        "ExpirationMonth": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 12
                        },
                        "Type": {
                          "type": "string",
                          "enum": [
                            "Visa",
                            "MasterCard",
                            "Discover",
                            "AmericanExpress",
                            "Unknown"
                          ]
                        }
                      },
                      "required": [
                        "AccountNumber",
                        "AccountHolder",
                        "Type",
                        "ExpirationMonth",
                        "ExpirationYear"
                      ]
                    },
                    "BankAccount": {
                      "type": "object",
                      "properties": {
                        "AccountNumber": {
                          "type": "string"
                        },
                        "RoutingNumber": {
                          "type": "string"
                        },
                        "AccountHolder": {
                          "type": "string"
                        },
                        "Type": {
                          "type": "string",
                          "enum": [
                            "Checking",
                            "Savings",
                            "MoneyMarket",
                            "Unknown"
                          ]
                        }
                      },
                      "required": [
                        "AccountNumber",
                        "RoutingNumber",
                        "AccountHolder",
                        "Type"
                      ]
                    }
                  },
                  "required": [
                    "Id",
                    "Active",
                    "Type"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer Account not found"
          }
        }
      }
    },
...
    "/customer_accounts/{id}/service_accounts": {
      "get": {
        "tags": [
          "customer_accounts",
          "service_accounts"
        ],
        "summary": "Retrieves service accounts for a customer id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to get service accounts for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/service_account"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No Service Accounts not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "customer_account": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Date_Of_Birth": {
            "type": "string",
            "format": "date"
          },
          "Email": {
            "type": "string",
            "format": "email"
          },
          "Prefix": {
            "type": "string"
          },
          "Suffix": {
            "type": "string"
          },
          "First_Name": {
            "type": "string"
          },
          "Middle_Name": {
            "type": "string"
          },
          "Last_Name": {
            "type": "string"
          },
          "Language": {
            "type": "string"
          },
          "Past_Due": {
            "type": "number"
          },
          "Bad_Debt_Balance": {
            "type": "number"
          },
          "Pledge_On_Entity": {
            "type": "boolean"
          },
          "Primary_Phone": {
            "type": "object",
            "properties": {
              "Type": {
                "type": "string",
                "enum": [
                  "Landline",
                  "Mobile"
                ]
              },
              "Area_Code": {
                "type": "string",
                "minLength": 3,
                "maxLength": 3,
                "pattern": "^\\d{3}$"
              },
              "Number": {
                "type": "string",
                "minLength": 7,
                "maxLength": 7,
                "pattern": "^\\d{7}$"
              },
              "Extension": {
                "type": "string"
              },
              "Agree_To_Marketing": {
                "type": "boolean"
              }
            },
            "required": [
              "Type",
              "Area_Code",
              "Number"
            ]
          },
          "EnrolledInShareAndSave": {
            "type": "boolean"
          }
        }
      },
      "customer_account_patch": {
        "type": "object",
        "properties": {
          "Email": {
            "type": "string",
            "format": "email"
          },
          "Primary_Phone_Number": {
            "type": "object",
            "properties": {
              "Type": {
                "type": "string",
                "enum": [
                  "Landline",
                  "Mobile"
                ]
              },
              "Area_Code": {
                "type": "string",
                "minLength": 3,
                "maxLength": 3,
                "pattern": "^\\d{3}$"
              },
              "Number": {
                "type": "string",
                "minLength": 7,
                "maxLength": 7,
                "pattern": "^\\d{7}$"
              },
              "Extension": {
                "type": "string"
              },
              "Agree_To_Marketing": {
                "type": "boolean"
              }
            },
            "required": [
              "Type",
              "Area_Code",
              "Number"
            ]
          }
        }
      },
      "customer_account_add_note_request": {
        "type": "object",
        "properties": {
          "Note": {
            "type": "string"
          },
          "Note_Type": {
            "type": "string",
            "enum": [
              "CC: ACCT NUMBER INQ",
              "CC: BALANCE INQUIRY",
              "CC: CUSTOMER COMPLAINT",
              "CC: EXTENSION GRANTED",
              "CC: INELIGIBLE FOR PYMT ARRG",
              "CC: NO BILL/DISC NOTICE RECVD",
              "CC: OTHER",
              "CC: PAYMENT CONFIRMATION",
              "CC: RECONNECT INQUIRY"
            ]
          },
          "Followup_Required": {
            "type": "boolean"
          },
          "Followup_Required_Date": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "Note",
          "Note_Type"
        ]
      },
...
      "error_list": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
...
      "pay_method": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "Active": {
            "type": "boolean"
          },
          "Type": {
            "type": "string",
            "enum": [
              "CreditCard",
              "eCheck",
              "Unknown"
            ]
          },
          "CreditCard": {
            "type": "object",
            "properties": {
              "AccountNumber": {
                "type": "string"
              },
              "AccountHolder": {
                "type": "string"
              },
              "ExpirationYear": {
                "type": "integer"
              },
              "ExpirationMonth": {
                "type": "integer",
                "minimum": 1,
                "maximum": 12
              },
              "Type": {
                "type": "string",
                "enum": [
                  "Visa",
                  "MasterCard",
                  "Discover",
                  "AmericanExpress",
                  "Unknown"
                ]
              }
            },
            "required": [
              "AccountNumber",
              "AccountHolder",
              "Type",
              "ExpirationMonth",
              "ExpirationYear"
            ]
          },
          "BankAccount": {
            "type": "object",
            "properties": {
              "AccountNumber": {
                "type": "string"
              },
              "RoutingNumber": {
                "type": "string"
              },
              "AccountHolder": {
                "type": "string"
              },
              "Type": {
                "type": "string",
                "enum": [
                  "Checking",
                  "Savings",
                  "MoneyMarket",
                  "Unknown"
                ]
              }
            },
            "required": [
              "AccountNumber",
              "RoutingNumber",
              "AccountHolder",
              "Type"
            ]
          }
        },
        "required": [
          "Id",
          "Active",
          "Type"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}